+++ resolwe_bio/tests/processes/test_jbrowse.py
@@ -14,6 +14,8 @@ class JbrowseProcessorTestCase(BioProcessTestCase):
         del refseq_track.output['refseq_track']['total_size']  # Non-deterministic output.
         self.assertFields(refseq_track, 'refseq_track', {'refs': ['seq'],
                                                          'file': 'seq/refSeqs.json'})
+        self.assertFields(refseq_track, 'species', 'Dictyostelium discoideum')
+        self.assertFields(refseq_track, 'build', 'dd-05-2009')
 
     @tag_process('jbrowse-gff3')
     def test_gff3_track(self):
@@ -25,6 +27,8 @@ class JbrowseProcessorTestCase(BioProcessTestCase):
 
         self.assertFields(gff, 'annotation_track', {'refs': ['tracks/annotation'],
                                                     'file': 'trackList.json'})
+        self.assertFields(gff, 'species', 'Dictyostelium discoideum')
+        self.assertFields(gff, 'build', 'dd-05-2009')
 
     @tag_process('jbrowse-gtf')
     def test_gtf_track(self):
@@ -35,11 +39,17 @@ class JbrowseProcessorTestCase(BioProcessTestCase):
         del gtf.output['annotation_track']['total_size']  # Non-deterministic output.
         self.assertFields(gtf, 'annotation_track', {'refs': ['tracks/annotation'],
                                                     'file': 'trackList.json'})
+        self.assertFields(gtf, 'species', 'Dictyostelium discoideum')
+        self.assertFields(gtf, 'build', 'dd-05-2009')
 
     @tag_process('jbrowse-bed')
     def test_bed_track(self):
         with self.preparation_stage():
-            inputs = {'src': 'bed_track.bed'}
+            inputs = {
+                'src': 'bed_track.bed',
+                'species': 'Dictyostelium discoideum',
+                'build': 'dd-05-2009'
+            }
             bed_file = self.run_process('upload-bed', inputs)
 
         bed = self.run_process('jbrowse-bed', {'bed': bed_file.pk})
@@ -47,6 +57,8 @@ class JbrowseProcessorTestCase(BioProcessTestCase):
 
         self.assertFields(bed, 'bed_track', {'refs': ['tracks/bed'],
                                              'file': 'trackList.json'})
+        self.assertFields(bed, 'species', 'Dictyostelium discoideum')
+        self.assertFields(bed, 'build', 'dd-05-2009')
 
     @tag_process('jbrowse-bam-coverage')
     def test_coverage_track(self):
@@ -60,6 +72,8 @@ class JbrowseProcessorTestCase(BioProcessTestCase):
 
         coverage = self.run_process('jbrowse-bam-coverage', {'bam': bam.pk})
         self.assertFile(coverage, 'bigwig_track', 'Jbrowse_genome_coverage.bw')
+        self.assertFields(coverage, 'species', 'Homo sapiens')
+        self.assertFields(coverage, 'build', 'hg19')
 
     @tag_process('jbrowse-bam-coverage-normalized')
     def test_norm_coverage_track(self):
@@ -74,3 +88,5 @@ class JbrowseProcessorTestCase(BioProcessTestCase):
         inputs = {'bam': bam.pk, 'size': 34000000}
         coverage = self.run_process('jbrowse-bam-coverage-normalized', inputs)
         self.assertFile(coverage, 'bigwig_track', 'Jbrowse_norm_genome_coverage.bw')
+        self.assertFields(coverage, 'species', 'Homo sapiens')
+        self.assertFields(coverage, 'build', 'hg19')